requirements.txt
aiogram==3.2.0
httpx==0.25.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
flask==3.0.0
gunicorn==21.2.0